major refactor of Recorder and its associated configuration
This feels like a big step in the right direction. It can possibly go further so that instead of a builder for the configuration the builder actually produces the Recorder instance / @Rule class. This at least gets the appropriate separation between config and lifecycle management so that all the extra config required for the proxy impl isn't getting in the way.