Created a failing test
So it fails playback when doing things concurrently using the
AsyncHttpClient from Apache Http Components.

This is similar to the behavior I found when I ran the proxy standalone.
Of note is that it doesn't seem to fail if the tape is READ_WRITE. Will
have to add a test to demonstrate that as well, then refactor all the
duplicated code out.