Fix ConfigurationSpec on Windows
ConfigurationSpec writes property files to a temporary tape root
directory, and then deletes them in the cleanup block. However, if the
call to propertiesFile.delete() fails, it simply returns false, rather
than throwing an exception.

On Microsoft Windows, the delete attempt usually fails, because it is
impossible in Windows to delete a memory-mapped file [1], which is what
the propertiesFile is, since Groovy's withWriter routine creates a
BufferedWriter. When the propertiesFile gets left behind, it later
causes the PostBodySpec to fail, which in turn causes the
SSLOverriderSpec to fail.

The traditional way to work around this bug is apparently [2] to call
System.gc() before attempting to delete the file. On some Windows
systems, this does not seem to work on the first attempt, but making
multiple attempts seems to work reliably.

[1] http://bugs.java.com/bugdatabase/view_bug.do?bug_id=4715154
[2] http://stackoverflow.com/questions/991489/i-cant-delete-a-file-in-java