Add option to enable authentication on the Betamax proxy.
Implements issue #174. The `ProxyConfiguration` now has a `proxyAuth()` method
that configures the username and password that the LittleProxy server will
require to accept connections. It's a convenience for any tool that uses Betamax
for testing and also wants to verify that its proxy support fully works. It
has no direct bearing on how Betamax itself operates.