Adding support for a quiet mode
This PR adds support for "quiet" mode. Currently, the READ_ONLY mode treats not finding a request as a hard error, throwing a NonWritableTapeException. This is certainly reasonable and desired behavior for many types of tests.

But for some tests, this behavior leads to excessive brittleness. If a code path under test is commonly used, this behavior means that any additional network calls require re-recording a significant number of tests even if the behavior is not expected to change.

In this PR, READ_ONLY_QUIET treats requests that are not present in the tape as simple 404 errors - similar to how the lack of a tape itself is presented as a 403. This allows tests to continue to pass without needing to be re-recorded.